Added function to download all JSON reports
Every scan has its own JSON report. The objective of this function
is to download all of the JSON reports of all the scans and save
them locally, in order to be able to extract all of the findings
and send them to an aggregator.

The reason of this has been to integrate Detectify to VulnWhisperer
(see more at https://github.com/austin-taylor/VulnWhisperer)